requirements.txt
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
pyrogram==2.0.106
tgcrypto==1.2.5
nest-asyncio==1.5.8
appdirs==1.4.4
filelock==3.14.0
beautifulsoup4==4.12.2
pytz==2024.1 
Pillow==10.4.0
